Move network mocking to separate module
