+++ app/src/App.test.tsx
@@ -1,6 +1,7 @@
 import { render, screen } from "@testing-library/react";
+import { testApiMock } from "./mocks.js";
 
-test("smoke test of accounts list via router", async () => {
+testApiMock("smoke test of accounts list via router", async () => {
   // We make our first request for /api_url while importing ApiClientContext, so
   // we need to delay importing this module until after test mocks have been set
   // up.
